Move regex into smaller checks and restrict ubuntu2004 tests
This commit will move a bigger regex check into smaller regex variables. Applying the commit will result in the order of the file to not matter, so long as the arguments are the same. This originally was not the case, and needed to be in a precise order.

Ubuntu2004 STIG guidelines for UBTU-20-010072 do not require that the parameter must not exist within common-auth if it already exists in /etc/security/faillock.conf. Additionally, there is no need to use "accounts" parameter as this is not needed per STIG guidelines.